Move GameGUI into Usercase Package
